+++ HotelBookingSystem.Core/Entities/Room.cs
@@ -9,11 +9,11 @@ namespace HotelBookingSystem.Domain.Entities
         public int RoomId { get; set; }
         public int HotelId { get; set; }
         public RoomType RoomType { get; set; }
-        public decimal PricePerNight { get; set; }
+        public double PricePerNight { get; set; }
         public int AdultCapacity { get; set; }
         public int ChildCapacity { get; set; }
         public bool FeaturedDeal { get; set; } 
-        public decimal? DiscountedPrice { get; set; }
+        public double? DiscountedPrice { get; set; }
         public Hotel Hotel { get; set; }
         public IList<string> ImagesUrl { get; set; }
         public ICollection<Booking> Bookings { get; set; }
